+++ tybotRoute/CHANGELOG.md
@@ -6,7 +6,7 @@ available on:
  ▶️ https://www.npmjs.com/package/@tiledesk/tiledesk-tybot-connector
 
 
-# v0.2.64 - rc
+# v0.2.64
 - removed system parameters from /reserved/parameters
 - Web Request Action: added multipart form-data support with files binary from URL source
 - Added DirAssistant Action